+++ appnav/src/main/kotlin/io/element/android/appnav/loggedin/LoggedInView.kt
@@ -19,13 +19,13 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.lifecycle.Lifecycle
 import io.element.android.appnav.R
 import io.element.android.libraries.architecture.AsyncData
-import io.element.android.libraries.designsystem.components.dialogs.ConfirmationDialog
 import io.element.android.libraries.designsystem.components.dialogs.ErrorDialog
 import io.element.android.libraries.designsystem.components.dialogs.ErrorDialogWithDoNotShowAgain
 import io.element.android.libraries.designsystem.preview.ElementPreview
 import io.element.android.libraries.designsystem.preview.PreviewsDayNight
 import io.element.android.libraries.designsystem.utils.OnLifecycleEvent
 import io.element.android.libraries.matrix.api.exception.isNetworkError
+import io.element.android.libraries.permissions.api.localnetwork.LocalNetworkPermissionDialogView
 import io.element.android.libraries.push.api.PusherRegistrationFailure
 import io.element.android.libraries.ui.strings.CommonStrings
 
@@ -84,38 +84,11 @@ fun LoggedInView(
         )
     }
 
-    LocalNetworkPermissionDialog(state = state)
-}
-
-@Composable
-private fun LocalNetworkPermissionDialog(state: LoggedInState) {
-    when (state.localNetworkPermissionDialog) {
-        LocalNetworkPermissionDialog.None -> Unit
-        LocalNetworkPermissionDialog.Rationale -> ConfirmationDialog(
-            title = stringResource(CommonStrings.screen_local_network_opt_in_title),
-            content = stringResource(CommonStrings.screen_local_network_opt_in_subtitle),
-            submitText = stringResource(CommonStrings.dialog_allow_access),
-            cancelText = stringResource(CommonStrings.action_not_now),
-            onSubmitClick = {
-                state.eventSink(LoggedInEvents.RequestLocationNetworkPermission)
-            },
-            onDismiss = {
-                state.eventSink(LoggedInEvents.DismissLocalNetworkPermissionPrompt)
-            },
-        )
-        LocalNetworkPermissionDialog.Settings -> ConfirmationDialog(
-            title = stringResource(CommonStrings.screen_local_network_opt_in_title),
-            content = stringResource(CommonStrings.screen_local_network_opt_in_subtitle),
-            submitText = stringResource(CommonStrings.action_open_settings),
-            cancelText = stringResource(CommonStrings.action_not_now),
-            onSubmitClick = {
-                state.eventSink(LoggedInEvents.RequestLocationNetworkPermission)
-            },
-            onDismiss = {
-                state.eventSink(LoggedInEvents.DismissLocalNetworkPermissionPrompt)
-            },
-        )
-    }
+    LocalNetworkPermissionDialogView(
+        dialog = state.localNetworkPermissionDialog,
+        onSubmit = { state.eventSink(LoggedInEvents.RequestLocationNetworkPermission) },
+        onDismiss = { state.eventSink(LoggedInEvents.DismissLocalNetworkPermissionPrompt) },
+    )
 }
 
 private fun Throwable.getReason(): String? {
