+++ libraries/permissions/api/src/main/kotlin/io/element/android/libraries/permissions/api/localnetwork/LocalNetworkPermissionAdvisor.kt
@@ -5,7 +5,7 @@
  * Please see LICENSE files in the repository root for full details.
  */
 
-package io.element.android.libraries.permissions.api
+package io.element.android.libraries.permissions.api.localnetwork
 
 interface LocalNetworkPermissionAdvisor {
     /**
