+++ libraries/permissions/impl/src/main/kotlin/io/element/android/libraries/permissions/impl/localnetwork/DefaultLocalNetworkPermissionAdvisor.kt
@@ -5,16 +5,16 @@
  * Please see LICENSE files in the repository root for full details.
  */
 
-package io.element.android.libraries.permissions.impl
+package io.element.android.libraries.permissions.impl.localnetwork
 
-import android.Manifest.permission.ACCESS_LOCAL_NETWORK
+import android.Manifest
 import android.os.Build
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.ContributesBinding
 import io.element.android.libraries.androidutils.network.LocalNetworkAddressClassifier
 import io.element.android.libraries.androidutils.network.LocalNetworkClassification
-import io.element.android.libraries.permissions.api.LocalNetworkPermissionAdvisor
 import io.element.android.libraries.permissions.api.PermissionStateProvider
+import io.element.android.libraries.permissions.api.localnetwork.LocalNetworkPermissionAdvisor
 import io.element.android.services.toolbox.api.sdk.BuildVersionSdkIntProvider
 
 @ContributesBinding(AppScope::class)
@@ -25,7 +25,7 @@ class DefaultLocalNetworkPermissionAdvisor(
 ) : LocalNetworkPermissionAdvisor {
     override suspend fun shouldRequestPermissionFor(homeserverUrl: String): Boolean {
         if (!buildVersionSdkIntProvider.isAtLeast(Build.VERSION_CODES.CINNAMON_BUN)) return false
-        if (permissionStateProvider.isPermissionGranted(ACCESS_LOCAL_NETWORK)) return false
+        if (permissionStateProvider.isPermissionGranted(Manifest.permission.ACCESS_LOCAL_NETWORK)) return false
         return when (classifier.classify(homeserverUrl)) {
             LocalNetworkClassification.LocalIp -> true
             LocalNetworkClassification.PublicIp,
